+++ src/Deedle/Indices/VirtualIndex.fs
@@ -27,15 +27,20 @@ open System.Collections.ObjectModel
 // VirtualOrderedIndex - index where the keys are defined by the specified sorted source
 // ------------------------------------------------------------------------------------------------
 
+/// Represents an ordered index based on data provided by a virtual source.
+/// The index can be used by BigDeedle virtual frames and series, without accessing
+/// all data from the data source.
+///
+/// The index only evaluates the full key collection when needed. Most of the actual
+/// work is delegated to the `IVirtualVectorSource<'K>` value passed in the constructor.
 type VirtualOrderedIndex<'K when 'K : equality>(source:IVirtualVectorSource<'K>) =
-  // TODO: Assert that the source is ordered
-  // TODO: Another assumption here is that the source contains no NA values
   let keyAtAddr i = source.ValueAt(i).Value
   let keys = Lazy.Create(fun () ->
     source.AddressOperations.Range
     |> Seq.mapl (fun idx addr -> keyAtAddr (Location.known(addr, idx)))
     |> ReadOnlyCollection.ofSeq )
 
+  /// Returns the underlying source associated with the index
   member x.Source = source
 
   /// Implements structural equality check against another index
@@ -49,31 +54,35 @@ type VirtualOrderedIndex<'K when 'K : equality>(source:IVirtualVectorSource<'K>)
   override index.GetHashCode() =
     (index :> IIndex<_>).Keys |> Seq.structuralHash
 
+  // The interface mostly delegates all work to the source
   interface IIndex<'K> with
-    member x.AddressingScheme = VirtualAddressingScheme() :> _
+    member x.AddressingScheme = VirtualAddressingScheme(source.AddressingSchemeID) :> _
     member x.AddressOperations = source.AddressOperations
     member x.KeyAt(addr) = keyAtAddr (Location.delayed(addr, source.AddressOperations))
     member x.AddressAt(offs) = source.AddressOperations.AddressOf offs
     member x.KeyCount = source.Length
     member x.IsEmpty = source.AddressOperations.Range |> Seq.isEmpty 
     member x.Builder = VirtualIndexBuilder.Instance :> _
+    member x.Keys = keys.Value 
+    member x.IsOrdered = true
+    member x.Comparer = Comparer<'K>.Default
+    
     member x.KeyRange = 
       keyAtAddr (Location.known(source.AddressOperations.FirstElement, 0L)),
       keyAtAddr (Location.delayed(source.AddressOperations.LastElement, source.AddressOperations))
     
-    member x.Keys = keys.Value 
     member x.KeySequence : seq<'K> = 
       source.AddressOperations.Range
       |> Seq.mapl (fun idx addr -> keyAtAddr (Location.known(addr, idx)))
+    
     member x.Mappings = 
       source.AddressOperations.Range
       |> Seq.mapl (fun idx addr -> KeyValuePair(keyAtAddr (Location.known(addr, idx)), addr))
-    member x.IsOrdered = true
-    member x.Comparer = Comparer<'K>.Default
 
     member x.Locate(key) = 
       let loc = source.LookupValue(key, Lookup.Exact, fun _ -> true)
       if loc.HasValue then snd loc.Value else Address.invalid
+
     member x.Lookup(key, semantics, check) = 
       source.LookupValue(key, semantics, Func<_, _>(check))
 
@@ -81,15 +90,21 @@ type VirtualOrderedIndex<'K when 'K : equality>(source:IVirtualVectorSource<'K>)
 // VirtualOrdinalIndex - ordinal index where the keys are specified by the given 'Ranges'
 // ------------------------------------------------------------------------------------------------
 
-/// 
+/// Represents an ordinal index based on addressing provided by a virtual source.
+/// The index can be used by BigDeedle virtual frames and series, without accessing
+/// all data from the data source.
 and VirtualOrdinalIndex(ranges:Ranges<int64>, source:IVirtualVectorSource) =
+  
+  /// Returns the ranges of data mapped in this index
   member x.Ranges = ranges
+  /// Returns the source that is used to identify the index.
   member x.Source = source
+
   interface IIndex<int64> with
-    member x.AddressingScheme = VirtualAddressingScheme() :> _
+    member x.AddressingScheme = VirtualAddressingScheme(source.AddressingSchemeID) :> _
     member x.AddressOperations = source.AddressOperations
-    member x.KeyAt(addr) = ranges |> Ranges.offsetOfKey (source.AddressOperations.OffsetOf(addr))
-    member x.AddressAt(offs) = ranges |> Ranges.keyAtOffset offs |> source.AddressOperations.AddressOf
+    member x.KeyAt(addr) = ranges |> Ranges.keyAtOffset (source.AddressOperations.OffsetOf(addr))
+    member x.AddressAt(offs) = source.AddressOperations.AddressOf offs
     member x.KeyCount = source.Length
     member x.IsEmpty = source.AddressOperations.Range |> Seq.isEmpty
     member x.Builder = VirtualIndexBuilder.Instance :> _
@@ -101,25 +116,32 @@ and VirtualOrdinalIndex(ranges:Ranges<int64>, source:IVirtualVectorSource) =
     member x.Comparer = Comparer<int64>.Default
     member x.Locate(key) = source.AddressOperations.AddressOf(Ranges.offsetOfKey key ranges)
     member x.Lookup(key, semantics, check) = 
-      failwith "LOokup!"
-      //Ranges.lookup key semantics (addressing.AddressOf >> check) ranges 
-      //|> OptionalValue.map (fun (key, addr) ->
-      //    key, addressing.AddressOf addr)
+      Ranges.lookup key semantics (fun k _ -> check(source.AddressOperations.AddressOf k)) ranges 
+      |> OptionalValue.map (fun (k, offs) -> k, source.AddressOperations.AddressOf k)
 
 // ------------------------------------------------------------------------------------------------
 // VirtualIndexBuilder - implements operations on VirtualOrderedIndex and VirtualOrdinalIndex
 // ------------------------------------------------------------------------------------------------
 
+/// Implements `IIndexBuilder` interface for BigDeedle. This directly implements operations that can 
+/// be implemented on virtual vectors (mostly merging, slicing) and for other operations, it calls
+/// ordinary `LinearIndexBuilder`. The resulting `VectorConstruction` corresponds to the addressing
+/// scheme of the returned index (i.e. if we return virtual, we expect to build virtual vector; if
+/// we materialize, the vector builder also has to materialize).
 and VirtualIndexBuilder() = 
   let baseBuilder = IndexBuilder.Instance
 
+  /// Create linear index from the keys of a given index
   let materializeIndex (index:IIndex<'T>) =
     Linear.LinearIndexBuilder.Instance.Create(index.Keys, Some index.IsOrdered)
 
   static let indexBuilder = VirtualIndexBuilder()
+
+  /// Returns an instance of the virtual index builder
   static member Instance = indexBuilder
 
   interface IIndexBuilder with
+    // The following are simply delegated to `LinearIndexBuilder`
     member x.Create(keys:seq<_>, ordered) = baseBuilder.Create(keys, ordered)      
     member x.Create(keys, ordered) = baseBuilder.Create(keys, ordered)
     member x.Recreate(index) = baseBuilder.Recreate(index)
@@ -129,30 +151,50 @@ and VirtualIndexBuilder() =
     member x.Union(sc1, sc2) = baseBuilder.Union(sc1, sc2)
     member x.Intersect(sc1, sc2) = baseBuilder.Intersect(sc1, sc2)
     member x.LookupLevel(sc, key) = baseBuilder.LookupLevel(sc, key)
-    member x.DropItem((index:IIndex<'K>, vector), key) = 
+    member x.DropItem((index:IIndex<_>, vector), key) = 
       baseBuilder.DropItem((index, vector), key)
     member x.Aggregate(index, aggregation, vector, selector) = 
       baseBuilder.Aggregate(index, aggregation, vector, selector)
     member x.Resample(chunkBuilder, index, keys, close, vect, selector) = 
       baseBuilder.Resample(chunkBuilder, index, keys, close, vect, selector)
 
-      
+    // VirtualVector.Select returns a VirtualVector and so projecting
+    // virtual index also needs to return virtual index
+    member x.Project(index:IIndex<'K>) = index
+
+    // Materialize the index asynchronously
+    member x.AsyncMaterialize( (index:IIndex<'K>, vector) ) = 
+      match index with
+      | :? VirtualOrdinalIndex 
+      | :? VirtualOrderedIndex<'K> ->
+          async.Return(materializeIndex index), vector
+      | _ -> 
+          baseBuilder.AsyncMaterialize((index, vector))
+          
+
+    // If the index is ordered, return it as it is. Otherwise, materialize and order      
     member x.OrderIndex( (index, vector) ) = 
       if index.IsOrdered then index, vector
       else baseBuilder.OrderIndex( (index, vector) )
     
-    member x.Merge(scs:list<SeriesConstruction<'K>>, transform) = 
+    // We can merge series as long as they all have `VirtualOrderedIndex` or as long as they
+    // all have `VirtualOrdinalIndex`. If they differ or they are fully evaluated, we call the
+    // base vector builder and fully materialize them.
+    member x.Merge<'K when 'K : equality>(scs:list<SeriesConstruction<'K>>, transform) = 
 
-      let (|OrderedSources|_|) : list<SeriesConstruction<_>> -> _ = 
+      /// Succeeds if all the specified indices are `VirtualOrderedIndex`
+      let (|OrderedSources|_|) : list<SeriesConstruction<'K>> -> _ = 
         List.tryChooseBy (function
           | :? VirtualOrderedIndex<'K> as idx, vec -> Some(idx.Source, vec)
           | _ -> None)
     
-      let (|OrdinalSources|_|) : list<SeriesConstruction<_>> -> _ = 
+      /// Succeeds if all the specified indices are `VirtualOrdinalIndex`
+      let (|OrdinalSources|_|) : list<SeriesConstruction<'K>> -> _ = 
         List.tryChooseBy (function
           | :? VirtualOrdinalIndex as idx, vec -> Some(idx.Ranges, idx.Source, vec)
           | _ -> None)
 
+      // Handle the different 3 cases - ordered, ordinal, mixed/other
       match scs with
       | OrderedSources sources ->
           let newSource = (fst sources.Head).MergeWith(sources.Tail |> List.map fst)
@@ -170,15 +212,17 @@ and VirtualIndexBuilder() =
           let newVector = seq { for _, _, v in sources -> v } |> Seq.reduce (fun a b -> Vectors.Append(a, b))
           unbox (VirtualOrdinalIndex(newRanges, newSource)), newVector
       | _ -> 
-          // Materialize virtual vectors
           let scs = 
             [ for index, vector in scs ->
                 match index with
-                | :? VirtualOrderedIndex<'K>
-                | :? VirtualOrdinalIndex -> materializeIndex index, vector
+                | :? VirtualOrderedIndex<'K> | :? VirtualOrdinalIndex -> materializeIndex index, vector
                 | _ -> index, vector ]
           baseBuilder.Merge(scs, transform)
-          
+
+
+    // Search the given vector for a given value and return an index that represents only
+    // part of the data matching the value. This works efficiently for `VirtualOrderedIndex`
+    // values that provide the `LookupRange` operation to do the search
     member x.Search((index:IIndex<'K>, vector), searchVector:IVector<'V>, searchValue) = 
       let searchVector = 
         match searchVector with
@@ -186,45 +230,40 @@ and VirtualIndexBuilder() =
         | _ -> searchVector
 
       match index, searchVector with
-      | (:? VirtualOrdinalIndex as index), (:? VirtualVector<'V> as searchVector) ->
-          failwith "TODO: Search!"
-          (*
-          let mapping = searchVector.Source.LookupRange(searchValue)
-          let count =
-            match mapping with
-            | VectorRange.Custom(iv) -> iv.Count
-            | VectorRange.Range(lo, hi) -> failwith "TODO: (Address.asInt64 hi) - (Address.asInt64 lo) + 1L"
-          let newIndex = VirtualOrdinalIndex(Ranges.Create [ 0L, count-1L ])
-          unbox<IIndex<'K>> newIndex, GetRange(vector, mapping)
-          *)
-
       | (:? VirtualOrderedIndex<'K> as index), (:? VirtualVector<'V> as searchVector) ->
           let mapping = searchVector.Source.LookupRange(searchValue)
           let newIndex = VirtualOrderedIndex(index.Source.GetSubVector(mapping))
           newIndex :> _, GetRange(vector, mapping)
-
       | _ ->
-          failwith "TODO: Search - search would cause materialization"
+          baseBuilder.Search((index, vector), searchVector, searchValue)
 
+
+    /// Replace the index with another index (whose keys are specified by a vector)
     member x.WithIndex(index1:IIndex<'K>, indexVector:IVector<'TNewKey>, vector) =
       match indexVector with
       | :? VirtualVector<'TNewKey> as indexVector ->
+          // Assumes that indexVector.Source is ordered (impossible to check)
+          // Assumes that indexVector.Source has no duplicates/NANs (impossible)
           let newIndex = VirtualOrderedIndex(indexVector.Source)
-          // TODO: Assert that indexVector.Source is ordered and can actually be used as an index?
-          // TODO: Assert that indexVector.Source has no duplicate keys and no NANs.... (which is impossible to do)
           upcast newIndex, vector
       | _ -> 
         baseBuilder.WithIndex(index1, indexVector, vector)
 
+
+    // This materializes the index, except when we want to reindex according to
+    // the same index (optimization) in which case we just return it
     member x.Reindex(index1:IIndex<'K>, index2:IIndex<'K>, semantics, vector, cond) = 
       match index1, index2 with
-      | (:? VirtualOrdinalIndex as index1), (:? VirtualOrdinalIndex as index2) when index1.Ranges = index2.Ranges -> vector           
+      | (:? VirtualOrdinalIndex as index1), (:? VirtualOrdinalIndex as index2) 
+          when index1.Ranges = index2.Ranges -> vector           
       | _ -> baseBuilder.Reindex(index1, index2, semantics, vector, cond)
 
+
+    // Apply the specified address range restriction - for `VirtualOrderedIndex` and
+    // `VirtualOrdinalIndex`, we directly delegate this to the source. Otherwise, materialzie.
     member x.GetAddressRange<'K when 'K : equality>((index, vector), range) = 
       match index with
       | :? VirtualOrderedIndex<'K> as index ->
-          // TODO: Range checks
           let newIndex = VirtualOrderedIndex(index.Source.GetSubVector(range))
           let newVector = Vectors.GetRange(vector, range)
           newIndex :> IIndex<'K>, newVector
@@ -234,30 +273,27 @@ and VirtualIndexBuilder() =
             range |> RangeRestriction.map (fun addr -> 
               Ranges.keyAtOffset (index.Source.AddressOperations.OffsetOf(addr)) index.Ranges)
           
-          // TODO: range checks
+          // We know that 'K = int64, but we don't know what is the
+          // generic type of the source (it could be anything)
           let newIndex = 
             { new IVirtualVectorSourceOperation<_> with 
-                member x.Invoke(source) = VirtualOrdinalIndex(Ranges.restrictRanges ordinalRestr index.Ranges, source.GetSubVector(range)) }
+                member x.Invoke(source) = 
+                  let newRanges = Ranges.restrictRanges ordinalRestr index.Ranges
+                  VirtualOrdinalIndex(newRanges, source.GetSubVector(range)) }
             |> index.Source.Invoke          
           let newVector = Vectors.GetRange(vector, range)
           unbox<IIndex<'K>> newIndex, newVector
       | _ -> 
           baseBuilder.GetAddressRange((index, vector), range)
           
-    member x.Project(index:IIndex<'K>) = 
-      index
 
-    member x.AsyncMaterialize( (index:IIndex<'K>, vector) ) = 
-      match index with
-      | :? VirtualOrdinalIndex 
-      | :? VirtualOrderedIndex<'K> ->
-          async.Return(materializeIndex index), vector
-      | _ -> 
-          baseBuilder.AsyncMaterialize((index, vector))
-          
+    // Get a range determined by the specified keys in the index. This is more complex than 
+    // `GetAddressRange` because we first need to find the addresses corresponding to the keys
+    // (or when we want exclusive boundaries, just after/before the keys).
     member x.GetRange<'K when 'K : equality>( (index, vector), (optLo:option<'K * _>, optHi:option<'K * _>)) = 
       match index with
       | :? VirtualOrderedIndex<'K> as index ->
+          /// Helper to Lookup the address of a key at the boundary
           let getRangeKey bound lookup = function
             | None -> Some bound
             | Some(k, beh) -> 
@@ -266,38 +302,44 @@ and VirtualIndexBuilder() =
                 | OptionalValue.Present(_, addr) -> Some addr
                 | _ -> None
 
+          // Get the addresses of the lower and upper boundary (handling exclusive lookups too)
           let loIdx = getRangeKey index.Source.AddressOperations.FirstElement Lookup.Greater optLo
           let hiIdx = getRangeKey index.Source.AddressOperations.LastElement Lookup.Smaller optHi
           match loIdx, hiIdx with
           | Some loIdx, Some hiIdx when loIdx <= hiIdx ->
+              // If the range was correct, return an address sub-range
               let newVector = Vectors.GetRange(vector, RangeRestriction.Fixed(loIdx, hiIdx))
               let newIndex = VirtualOrderedIndex(index.Source.GetSubVector(RangeRestriction.Fixed(loIdx, hiIdx)))
               newIndex :> IIndex<'K>, newVector
           | _ ->
+              // Otherwise (e.g. lo > hi) return empty series
               Index.ofKeys [] :> _, VectorConstruction.Empty(0L)
 
       | :? VirtualOrdinalIndex as ordIndex & (:? IIndex<int64> as index) -> 
+          /// Helper that gets the key (in the source index) corresponding the the place where the
+          /// subrange actually starts (the keys in the index may not be continuous e.g. when 
+          /// we merge them) and so for non-inclusive lookups, we inc/dec the *offset*.
           let getRangeKey proj next = function
             | None -> proj index.KeyRange 
             | Some(k, BoundaryBehavior.Inclusive) -> unbox<int64> k
             | Some(k, BoundaryBehavior.Exclusive) -> 
                 Ranges.keyAtOffset (next (Ranges.offsetOfKey (unbox<int64> k) ordIndex.Ranges)) ordIndex.Ranges
-          let loKey, hiKey = getRangeKey fst ((+) 1L) optLo, getRangeKey snd ((-) 1L) optHi
-          let loAddr, hiAddr = 
-            // TODO: Too many silly conversions there and back again
-            ordIndex.Source.AddressOperations.AddressOf (Ranges.offsetOfKey loKey ordIndex.Ranges),
-            ordIndex.Source.AddressOperations.AddressOf (Ranges.offsetOfKey hiKey ordIndex.Ranges)
 
-          // TODO: range checks
-          let restr = RangeRestriction.Fixed(loKey, hiKey)
-          let range = RangeRestriction.Fixed(loAddr, hiAddr)
-          let newVector = Vectors.GetRange(vector, range)
+          // Get the addresses, make range restriction
+          let loKey, hiKey = getRangeKey fst ((+) 1L) optLo, getRangeKey snd ((-) 1L) optHi
+          let loAddr, hiAddr = index.Locate(loKey), index.Locate(hiKey)
+          let keyRestr = RangeRestriction.Fixed(loKey, hiKey)
+          let addrRestr = RangeRestriction.Fixed(loAddr, hiAddr)
+          
+          // Make new vector and new index
+          let newVector = Vectors.GetRange(vector, addrRestr)
           let newSource =
             { new IVirtualVectorSourceOperation<_> with
-                member x.Invoke(source) = source.GetSubVector(range) :> IVirtualVectorSource }
+                member x.Invoke(source) = source.GetSubVector(addrRestr) :> IVirtualVectorSource }
             |> ordIndex.Source.Invoke
-          let newIndex = VirtualOrdinalIndex(Ranges.restrictRanges restr ordIndex.Ranges , newSource) 
+          let newIndex = VirtualOrdinalIndex(Ranges.restrictRanges keyRestr ordIndex.Ranges , newSource) 
           unbox<IIndex<'K>> newIndex, newVector
           
       | _ -> 
+          // Otherwise, materialize the index and perform slicing on it
           baseBuilder.GetRange((index, vector), (optLo, optHi))